git-svn-id: https://subversion.imagemagick.org/subversion/ImageMagick/trunk@1789 aa41f4f7-0bf4-0310-aa73-e5a19afd5a74

+++ coders/png.c
@@ -2690,9 +2690,9 @@ static Image *ReadOnePNGImage(MngInfo *mng_info,
         {
           indice=(IndexPacket) (*r++);
           indexes[x]=indice;
-          q->red=image->colormap[indice].red;
-          q->green=image->colormap[indice].green;
-          q->blue=image->colormap[indice].blue;
+          q->red=image->colormap[(int) indice].red;
+          q->green=image->colormap[(int) indice].green;
+          q->blue=image->colormap[(int) indice].blue;
         }
         if (SyncAuthenticPixels(image,exception) == MagickFalse)
           break;
